Fix parsing of URL into keyword arguments for connection creation (#8)
The `connect` method accepts a preformatted URL instead of the
connection parameters as individual arguments. However, in that case SSL
options are parsed from the URL and parsed as a normal dictionary to to
connection factory of `pika`, but this has long been deprecated. It is
now necessary to create a `pika.SSLOptions` object and pass that in.

A notable parameter that is not parsed is the `ssl_version`, which means
that the default `PROTOCOL_TLS` will be used when the context is
created. This default value lets the server decide what protocol should
be used and provides the most compatibility with other versions.